Move brush files
