+++ src/tools/behaviors/brushPaintMessages.js
